adding authentication and authorization to the service, adding patching for the progress, auto calculator for the status

+++ GoaService/src/main/java/com/GoalService/GoalService/domain/UpdateGoalResponse.java
@@ -14,7 +14,7 @@ public class UpdateGoalResponse {
     private String title;
     private String description;
     private LocalDate targetDate;
-    private long progress;
+    private int progress;
     private Category category;
     private Status status;
 }
